# Conflicts:
#	src/main/java/com/hear/hear/Controllers/AdminController.java
#	src/main/java/com/hear/hear/Controllers/FacultyMemberController.java
#	src/main/java/com/hear/hear/Controllers/StudentController.java
#	src/main/java/com/hear/hear/Controllers/UserController.java
#	src/main/java/com/hear/hear/Mappers/FacultyMemberMapping.java
#	src/main/java/com/hear/hear/Mappers/FavouriteMaterialsMapping.java
#	src/main/java/com/hear/hear/Mappers/RegisterUserRequest.java
#	src/main/java/com/hear/hear/Mappers/StudentMapping.java
#	src/main/java/com/hear/hear/Mappers/UserAllScheduleMapping.java
#	src/main/java/com/hear/hear/Mappers/UserMapping.java
#	src/main/java/com/hear/hear/Repositories/ClassRepository.java
#	src/main/java/com/hear/hear/Repositories/FacultyMemRepository.java
#	src/main/java/com/hear/hear/Repositories/StudentRepository.java
#	src/main/java/com/hear/hear/Repositories/UserRepository.java
#	src/main/java/com/hear/hear/authentication/AuthController.java
#	src/main/java/com/hear/hear/authentication/AuthService.java
#	src/main/java/com/hear/hear/authentication/JwtService.java
#	src/main/java/com/hear/hear/dtos/FacultyProfileDto.java
#	src/main/java/com/hear/hear/dtos/FavMaterials.java
#	src/main/java/com/hear/hear/dtos/StudentProfileDto.java
#	src/main/java/com/hear/hear/entities/Materials.java
#	src/main/java/com/hear/hear/entities/Notification.java
#	src/main/java/com/hear/hear/entities/User.java
#	src/main/java/com/hear/hear/filters/LoggingFilter.java
#	src/main/java/com/hear/hear/services/ClassService.java
#	src/main/java/com/hear/hear/services/FacultyMemberService.java
#	src/main/java/com/hear/hear/services/UserService.java

+++ src/main/java/com/hear/hear/authentication/Jwt.java
@@ -18,8 +18,8 @@ public class Jwt {
         return claims.getExpiration().before(new Date());
     }
 
-    public Long getUserId() {
-        return Long.valueOf(claims.getSubject());
+    public Integer getUserId() {
+        return Integer.valueOf(claims.getSubject());
     }
 
     public Role getRole() {
@@ -29,4 +29,6 @@ public class Jwt {
     public String toString() {
         return Jwts.builder().claims(claims).signWith(secretKey).compact();
     }
+
+
 }
